Pixel Types
Created a WorldDataHandler which will be a readonly container that holds the types of pixels in the game. Every PixelType (Below) can have up to currently 10 colours, these colours can be used as a lookup table to find out behaviours and quantifiers. I'm not sold that this is the best approach, but seems fairly easy to pull off.
Started adding pixel types
- BasePixel
- GroundPixel
- WaterPixel
Base Pixel contains requirements, and will in time likely have an update behaviour that pixels modify. (Water, Solid, Gas, etc)