New ShaderManager
Created a Shader manager which will load, compile and store references to shaders. These shaders can be enabled through the Manager or by asking for the reference back and doing it normally.

This class is a singleton, it should remain safe so long as the objects using it aren't trying to reference uncompiled shaders.