New texture format
Instead of each chunk holding the colour/texture of its chunk. We how onto an Index that represents what kind of pixel each pixel is. This allows us to use 6 bits for type, and 26 bits for other data.

This is good as it means when moving data around on cpu we don't have to hand-hold multiple variables any time a pixel moves.

Bit of a messy commit, but is currently functional. Draw times are reduced a bit, updates appear to be a fair bit faster as well now that only 1 int has to be moved around representing the pixel instead of colour + data.